requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Login>=0.5
Flask-WTF>=1.0
Flask-Migrate>=3.0
Config
WTForms>=2.3
Werkzeug>=2.0
Jinja2>=3.0
itsdangerous>=2.0
python-dotenv>=0.19

# For AI and monitoring
opencv-python>=4.5
DeepFace>=0.0.79
fer>=22.4.0
librosa>=0.9
pynput>=1.7
mouseinfo>=0.1.3
tf-keras>=2.6

# For Gemini API (if using Google API client)
google-api-python-client>=2.0
google-generativeai>=0.1.0

# For JSON and HTTP
requests>=2.25

# For threading and background tasks (standard library, but include for clarity)
thread6

# For real-time WebSocket support
Flask-SocketIO>=5.0
# eventlet is recommended for Flask-SocketIO
eventlet>=0.33 